worked out what could be the appropriate way of handleing events in backbone.
the view watches for the event and then does something when the model changes.

I don't know if I like the fact that i have two methods one to toggle the change and then one to watch for the event....  but it probably will save me later to watch for the event.